API/requirements.txt
fastapi 
uvicorn 
mysql-connector-python 
bcrypt 
python-dotenv 
aiosmtplib